1st commit - Seperates data methods from display methods. Optimises displaying normals so calculation of them is only performed when they change
